feat(looper): widen setTrackLoopRegion frame contract to int64
Loop-region start/end frames now cross the JNI boundary as int64
(Kotlin Long / jlong / C++ int64_t) so the ABI never has to widen when
high-tier devices record multi-hour tracks (plan §3.5). TrackBuffer
still stores frames as int32, so AudioLooper saturates the region into
int32 range before delegating (a valid region is well under INT_MAX
frames ≈ 12.4 h @ 48 kHz). Adds <cstdint> for the saturation bounds.

Covered by additions in test_audio_looper.cpp.

+++ audio/src/androidMain/kotlin/com/watermellonstudios/audio/internal/bridge/AudioNativeBridge.kt
@@ -2699,7 +2699,7 @@ class AudioNativeBridge private constructor() : IAudioNativeBridge {
     private external fun nativeLooperGetDroppedEvents(): Long
     private external fun nativeLooperSetMasterVolume(volume: Float)
     private external fun nativeLooperGetMasterVolume(): Float
-    private external fun nativeLooperSetTrackLoopRegion(trackIndex: Int, startFrame: Int, endFrame: Int)
+    private external fun nativeLooperSetTrackLoopRegion(trackIndex: Int, startFrame: Long, endFrame: Long)
     private external fun nativeLooperResetTrackLoopRegion(trackIndex: Int)
     private external fun nativeLooperGetTrackLoopStart(trackIndex: Int): Int
     private external fun nativeLooperGetTrackLoopEnd(trackIndex: Int): Int
@@ -2968,7 +2968,7 @@ class AudioNativeBridge private constructor() : IAudioNativeBridge {
     fun looperGetMasterVolume(): Float = nativeLooperGetMasterVolume()
 
     // Loop region (lock-free)
-    fun looperSetTrackLoopRegion(trackIndex: Int, startFrame: Int, endFrame: Int) =
+    fun looperSetTrackLoopRegion(trackIndex: Int, startFrame: Long, endFrame: Long) =
         nativeLooperSetTrackLoopRegion(trackIndex, startFrame, endFrame)
     fun looperResetTrackLoopRegion(trackIndex: Int) = nativeLooperResetTrackLoopRegion(trackIndex)
     fun looperGetTrackLoopStart(trackIndex: Int): Int = nativeLooperGetTrackLoopStart(trackIndex)
